code/requirements.txt
appdirs==1.4.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
Brotli==1.0.9
carla==0.9.14
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.5
cryptography==41.0.3
cuda-python==11.8.2
cycler==0.11.0
Cython==3.0.0
cytoolz==0.12.0
dask==2023.4.1
docopt==0.6.2
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.25.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.35
glfw==2.6.2
glumpy==1.2.0
huggingface-hub==0.19.0
humanfriendly==10.0
idna==3.4
imagecodecs==2023.1.23
imageio==2.31.4
importlib-metadata==6.0.0
importlib-resources==6.1.0
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
Js2Py==0.74
kiwisolver==1.4.4
kornia==0.6.12
locket==1.0.0
lpips==0.1.4
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.7.1
mpmath==1.3.0
munkres==1.1.4
networkx==3.1
numpy==1.23.5
nvidia-pyindex==1.0.9
onnx==1.15.0
onnx-graphsurgeon==0.3.12
onnxconverter-common==1.14.0
onnxruntime==1.16.1
onnxruntime-gpu==1.16.1
onnxscript==0.1.0.dev20231102
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
partd==1.4.1
pascal-voc-writer==0.1.4
Pillow==10.0.0
pipwin==0.5.2
platformdirs==3.10.0
ply==3.11
pooch==1.7.0
prefetch-generator==1.0.3
protobuf==3.20.2
psutil==5.9.5
py-cpuinfo==9.0.0
pycparser==2.21
pygame==2.5.0
pyglet==2.0.8
pyjsparser==2.7.1
PyOpenGL==3.1.7
pyOpenSSL==23.2.0
pyparsing==3.0.9
pypiwin32==223
PyPrind==2.11.3
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyreadline3==3.4.1
PySDL2==0.9.16
pysdl2-dll==2.28.0
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.2
pytools==2023.1
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.0
SciencePlots==2.1.0
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
setuptools==68.0.0
sip==6.7.12
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
sympy==1.11.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.4.12
timm==0.9.10
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
triangle==20220202
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uff==0.6.9
ultralytics==8.0.175
urllib3==1.26.18
wheel==0.41.2
win-inet-pton==1.1.0
yacs==0.1.8
zipp==3.11.0
